Make sorting adapt to data type + tmp plotting l.v_gene
Having the -1 in the name required the natural sorting to be reversed.
Now we check for the - and reverse automatically if needed.

Plotting vor l.v_gene has been enabled if the chain parameter is neither
l or h. Chain parameter in this method needs to be updated to accept
any column added by Read10X_vdj()