Generalise naming from (h)eavy/(l)ight to vdj/vj
When both TCR and BCR were loaded in the Seurat-object, this resulted in
~80 metadata columns. This was due to the naming of the metadata columns
h./v. (BCR) vs a./b. (TCR).
By renaming to a more general vdj./vj., the metadata is now a lot less
crowded and there should be no more confusion about which VDJ assay is
currently active.