Load sequence information from airr file when available
The airr file contains some extra information, such as sequences and
start/end positions, that is not available in the annotation csv.
This change also added the quiet parameter to the load function so the
warning can be ignored. This is useful for the tests, since the test
output is otherwise full of warning messages